requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
xlsxwriter>=3.0.0
pydantic>=2.0.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional dependencies for interface testing
requests>=2.28.0